+++ src/commands/redeploy.ts
@@ -4,7 +4,7 @@ import { SpinnerMsg } from "./command";
 
 export function redeploy(node: ClabLabTreeNode) {
   const spinnerMessages: SpinnerMsg = {
-    progressMsg: "Redeploying Lab... Check the Output window (Containerlab) for detailed progress.",
+    progressMsg: "Redeploying Lab... ",
     successMsg: "Lab redeployed successfully!"
   };
   const redeployCmd = new ClabCommand("redeploy", node, spinnerMessages);
@@ -13,7 +13,7 @@ export function redeploy(node: ClabLabTreeNode) {
 
 export function redeployCleanup(node: ClabLabTreeNode) {
   const spinnerMessages: SpinnerMsg = {
-    progressMsg: "Redeploying Lab (cleanup)... Check the Output window (Containerlab) for detailed progress.",
+    progressMsg: "Redeploying Lab (cleanup)... ",
     successMsg: "Lab redeployed (cleanup) successfully!"
   };
   const redeployCmd = new ClabCommand("redeploy", node, spinnerMessages);